+++ commands/say.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const Settings = require('../utils/Settings');
+const defaultVoice = Settings.getSetting( __filename, 'defaultVoice' );
 const Say = require('../utils/Say');
 const regex = new RegExp( /^(!|\/)say\s(.+)$/ );
 
@@ -15,7 +17,7 @@ module.exports = [{
 		message = message.substr( 0, 50 );
 
         // Allow users to override the voice
-        var voice = 'Victoria';
+        var voice = defaultVoice;
         var match = /^\-voice\s(\w+)\s(.+)/.exec( message );
         if ( match ) {
             voice = match[1];