+++ commands/user-status.js
@@ -1,6 +1,8 @@
 'use strict';
 
 const runtime = require('../utils/Runtime');
+const Settings = require('../utils/Settings');
+const availableStatuses = Settings.getSetting( __filename, 'statuses' );
 const setStatusRegex = new RegExp( /^(!|\/)setstatus\s(.+)\s(\w+)/ );
 const getStatusRegex = new RegExp( /^(!|\/)getstatus\s(.+)/ );
 
@@ -11,12 +13,8 @@ module.exports = [{
     types: ['message'],
     regex: /^(!|\/)status/,
     action: function( chat, stanza ) {
-		var status = stanza.user.status;
-		if ( !status ) {
-			status = 'Viewer';
-		}
-
-		chat.sendMessage(`${stanza.user.username} is set to: ${status}`);
+		let statusObj = availableStatuses[ stanza.user.status ];
+		chat.sendMessage(`${stanza.user.username} is set to: ${ statusObj.title }`);
     }
 }, {
 	name: '!getstatus {username}',
@@ -39,12 +37,8 @@ module.exports = [{
 			return;
 		}
 
-		var status = user.status;
-		if ( !status ) {
-			status = 'Viewer';
-		}
-
-		chat.sendMessage(`${username} is set to: ${status}`);
+		let statusObj = availableStatuses[ user.status ];
+		chat.sendMessage(`${username} is set to: ${ statusObj.title }`);
     }
 }, {
 	name: '!setstatus {username} {status}',
@@ -60,6 +54,12 @@ module.exports = [{
 				username = username.substr(1);
 			}
 
+			// Check if the status to set is an available status
+			if ( !isAvailableStatus( statusToSet ) ) {
+				chat.replyTo( username, `${ statusToSet } is not a valid status.` );
+				return;
+			}
+
 			// Look up the user
 			var users = runtime.brain.get( 'users' ) || {};
 			var user = users[ username ];
@@ -73,7 +73,22 @@ module.exports = [{
 			user.status = statusToSet;
 			runtime.brain.set('users', users);
 
-			chat.replyTo(username, `is now a ${statusToSet}!` );
+			let statusObj = availableStatuses[ statusToSet ];
+			chat.replyTo(username, `is now a ${statusObj.title}!` );
 		}
     }
 }];
+
+/**
+ * Returns a boolean if the passed-in status is
+ * an available status.
+ * @param  {String}  status
+ * @return {Boolean}
+ */
+function isAvailableStatus( statusID ) {
+	statusID = statusID.toLowerCase();
+	let statuses = Object.keys( availableStatuses ).map( ( statusKey ) => {
+		return statusKey.toLowerCase();
+	} );
+	return statuses.indexOf( statusID ) >= 0;
+}